change output type from long to int

+++ include/wav2wav.h
@@ -115,7 +115,7 @@ private:
 };
 
 
-std::tuple<std::vector<long>, int, tensor_info<float>, tensor_info<float>>
+std::tuple<std::vector<int>, int, tensor_info<float>, tensor_info<float>>
 next_token_A1T2(ONNXModel &gpt, tensor_info<float> &input_embs_concat, tensor_info<long> &input_pos_tensor,
                 tensor_info<float> &past_ks_tensor, tensor_info<float> &past_vs_tensor, int sub_step, float temperature,
                 int top_k, float top_p);
